adding updated dynamic scenic example
