+++ examples/carla/scenic/carlaChallenge2.scenic
@@ -1,0 +1,57 @@
+""" Scenario Description
+Traffic Scenario 02.
+Longitudinal control after leading vehicle’s brake.
+The leading vehicle decelerates suddenly due to an obstacle and the ego-vehicle must perform an
+emergency brake or an avoidance maneuver.
+"""
+
+## SET MAP AND MODEL (i.e. definitions of all referenceable vehicle types, road library, etc)
+param map = localPath('./Town01.xodr')  # or other CARLA map that definitely works
+param carla_map = 'Town01'
+model scenic.simulators.carla.model
+
+## CONSTANTS
+EGO_MODEL = "vehicle.lincoln.mkz2017"
+EGO_SPEED = 10
+EGO_BRAKING_THRESHOLD = 12
+
+LEAD_CAR_SPEED = 10
+LEADCAR_BRAKING_THRESHOLD = 10
+
+BRAKE_ACTION = 1.0
+
+## DEFINING BEHAVIORS
+# EGO BEHAVIOR: Follow lane, and brake after passing a threshold distance to the leading car
+behavior EgoBehavior(speed=10):
+    try:
+        do FollowLaneBehavior(speed)
+
+    interrupt when withinDistanceToAnyCars(self, EGO_BRAKING_THRESHOLD):
+        take SetBrakeAction(BRAKE_ACTION)
+
+# LEAD CAR BEHAVIOR: Follow lane, and brake after passing a threshold distance to obstacle
+behavior LeadingCarBehavior(speed=10):
+    try: 
+        do FollowLaneBehavior(speed)
+
+    interrupt when withinDistanceToAnyObjs(self, LEADCAR_BRAKING_THRESHOLD):
+        take SetBrakeAction(BRAKE_ACTION)
+
+## DEFINING SPATIAL RELATIONS
+# Please refer to scenic/domains/driving/roads.py how to access detailed road infrastructure
+# 'network' is the 'class Network' object in roads.py
+
+# make sure to put '*' to uniformly randomly select from all elements of the list, 'lanes'
+lane = Uniform(*network.lanes)
+
+obstacle = Trash on lane.centerline
+
+leadCar = Car following roadDirection from obstacle for Range(-50, -30),
+        with behavior LeadingCarBehavior(LEAD_CAR_SPEED)
+
+ego = Car following roadDirection from leadCar for Range(-15, -10),
+        with blueprint EGO_MODEL,
+        with behavior EgoBehavior(EGO_SPEED)
+
+require (distance to intersection) > 80
+terminate when ego.speed < 0.1 and (distance to obstacle) < 30
